src/requirements.txt
django==2.1.1
django-cors-headers
django-filter
djangorestframework
gunicorn
meld3
pytz
sqlparse
